showing file size alongsize name + if file is not found at source, showing the entry in red. major reliability issue fixed

+++ app/src/main/java/dev/sanskar/fileboi/models/FileMetadata.java
@@ -2,14 +2,14 @@ package dev.sanskar.fileboi.models;
 
 public class FileMetadata {
 
-    private int sizeInBytes;
+    private long sizeInBytes;
     private String lastModified;
 
-    public int getSizeInBytes() {
+    public long getSizeInBytes() {
         return sizeInBytes;
     }
 
-    public void setSizeInBytes(int sizeInBytes) {
+    public void setSizeInBytes(long sizeInBytes) {
         this.sizeInBytes = sizeInBytes;
     }
